+++ domain/src/main/java/ru/sandbox/swa/model/WeatherItem.kt
@@ -1,8 +1,8 @@
 package ru.sandbox.swa.model
 
 data class WeatherItem(
-    val id: Int = 0,
-    val date: Int = 0,
+    val id: Long = 0L,
+    val date: Long = 0L,
     val temperature: Double = 0.0,
     val humidity: Int = 0,
     val windSpeed: Double = 0.0
